# docs/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API: que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año",
    "version": "1.0.0"
  },
  "paths": {
    "/api/grupos/:grupoId/recaudacion": {
      "get": {
        "summary": "Obtiene el total de recaudación del grupo para el año seleccionado. Usado para refrescar las tarjetas superiores sin F5.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/grupos/:grupoId/pagos": {
      "post": {
        "summary": "Registra un nuevo pago en el sistema. Retorna el pago creado y el total de recaudación actualizado para el año.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "summary": "Lista todos los pagos del grupo filtrados por año. Soporta paginación y búsqueda.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/grupos/:grupoId/resumen-anual": {
      "get": {
        "summary": "Obtiene el resumen de recaudación por año. Usado para mostrar las tarjetas superiores con totales por año.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/grupos/:grupoId/pagos/:pagoId": {
      "put": {
        "summary": "Actualiza un pago existente y retorna la recaudación actualizada del año.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          },
          {
            "name": "pagoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del pago a actualizar"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Elimina un pago y retorna la recaudación actualizada del año.",
        "tags": [
          "que, al registrar un nuevo pago con éxito desde la pestaña Pagos, el total «Recaudación» de las tarjetas superiores se refresque en la misma sesión sin F5 ni cambiar manualmente el año"
        ],
        "parameters": [
          {
            "name": "grupoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del grupo deportivo"
          },
          {
            "name": "pagoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del pago a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}